lots of small tweaks and optimizations

+++ src/Ppu.h
@@ -12,10 +12,10 @@ struct scanline_prerender {};
 struct scanline_postrender {};
 
 struct scanline_render {
-	explicit scanline_render(uint16_t *p)
+	explicit scanline_render(uint32_t *p)
 		: buffer(p) {
 	}
-	uint16_t *const buffer;
+	uint32_t *const buffer;
 };
 
 namespace nes::ppu {
